+++ lopatina.sofia/S2/calculate_postfix.hpp
@@ -1,0 +1,12 @@
+#ifndef CALCULATE_POSTFIX_HPP
+#define CALCULATE_POSTFIX_HPP
+
+#include "queue.hpp"
+#include "types.hpp"
+
+namespace lopatina
+{
+  long long calculatePostfix(Queue< PostfixType > & queue_postfix);
+}
+
+#endif
